+++ pkg/commands/logging/https/create.go
@@ -39,6 +39,7 @@ type CreateCommand struct {
 	JSONFormat        argparser.OptionalString
 	MessageType       argparser.OptionalString
 	Method            argparser.OptionalString
+	Period            argparser.OptionalInt
 	Placement         argparser.OptionalString
 	ProcessingRegion  argparser.OptionalString
 	RequestMaxBytes   argparser.OptionalInt
@@ -84,6 +85,7 @@ func NewCreateCommand(parent argparser.Registerer, g *global.Data) *CreateComman
 	c.CmdClause.Flag("json-format", "Enforces valid JSON formatting for log entries. Can be disabled 0, array of json (wraps JSON log batches in an array) 1, or newline delimited json (places each JSON log entry onto a new line in a batch) 2").Action(c.JSONFormat.Set).StringVar(&c.JSONFormat.Value)
 	common.MessageType(c.CmdClause, &c.MessageType)
 	c.CmdClause.Flag("method", "HTTP method used for request. Can be POST or PUT. Defaults to POST if not specified").Action(c.Method.Set).StringVar(&c.Method.Value)
+	common.Period(c.CmdClause, &c.Period)
 	common.Placement(c.CmdClause, &c.Placement)
 	common.ProcessingRegion(c.CmdClause, &c.ProcessingRegion, "HTTPS")
 	c.CmdClause.Flag("request-max-bytes", "Maximum size of log batch, if non-zero. Defaults to 100MB").Action(c.RequestMaxBytes.Set).IntVar(&c.RequestMaxBytes.Value)
@@ -191,6 +193,9 @@ func (c *CreateCommand) ConstructInput(serviceID string, serviceVersion int) (*f
 		input.ResponseCondition = &c.ResponseCondition.Value
 	}
 
+	if c.Period.WasSet {
+		input.Period = &c.Period.Value
+	}
 	if c.Placement.WasSet {
 		input.Placement = &c.Placement.Value
 	}