+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Enhancements:
 - feat(logging): Add support for 'CompressionCodec' and 'GzipLevel' attribute to the HTTPS endpoint.
 - feat(kvstoreentry): Add support for the 'prefix' parameter for List operations ([#1526](https://github.com/fastly/cli/pull/1526))
+- feat(logging): Add support for 'Period' attribute to the HTTPS endpoint.
 
 ### Bug fixes: